test(KEN-720): a package-managed install is left alone [no-changelog]
The early return that tells a Homebrew or AUR install to use its own
package manager had no test, and none was possible: the channel came from
the process's real current_exe, which nothing here can put under /usr or a
brew prefix. Deleting the whole block left every kendex-cli suite green.

run now reads the process and hands what it read to run_on, which is the
whole update. A test gives it a channel, a temp command on disk, and a
file:// feed offering a newer release, then reads the bytes back. Managed
leaves them; Direct replaces them. Both mutations go red: dropping the
early return alone turns the managed run into an error, and dropping the
core refusal with it overwrites the command the package manager owns.

The tests moved to a sibling module, which is where the size ratchet's
seam is for this file.
